[eta] Properly handle comm in stat fields
The comm field in /proc/*/stat is enclosed by parentheses to guard
against commands with spaces in the name (I presume), but you can
easily craft a command that has a closing parenthesis in the name.
So let's just ask the kernel what the command name is via /proc/*/comm
and replace that with a placeholder - we don't need the command name
field anyway.