spothole: Clean things up when there's no running stream
The old code depended on at least one Spotify stream being in a RUNNING
state - this is all well and good if your music listening is roughly
balanced among instances, since while you could still build up a number of
dangling clients, listening to music on an instance will eventually
allow spothole to do its job.  But if, for some reason, there's an
instance that isn't used to listen for a while, the dangling clients can
build up enough where you can't even *get* a running stream, so you'll
be staring at Spotify, wondering why it can't play music, until you
remember why you wrote spothole in the first place, and restart Spotify.

Apparently the way the Spotify desktop client works on Linux is
something like this: an instance will create a new client when that
instance becomes "focused" - ie. it will be where music is played when
the user next asks to play music.  The focus can change due to manual
user intervention, but also if the currently focused instance drops away
- eg. if I was playing music from my phone, stopped, and Android decided
to suspend the Spotify app or something.  The newly created client isn't
referenced by any streams, but it can *become* referenced the next time
that instance plays music.

AFAICT, the only clients that ever get used to play music are 1) the
client attached to the stream with the maximum index, and 2) the client
with the maximum index.  So feel free to clean up any *other* client you
may find.