Fix man-pager to properly restore window title upon exit
Vim tries to restore the terminal title upon exit, but *apparently* this
doesn't work the best under urxvt when vim is reading from standard
input.

So stuff the rendered man page into a temporary file (using zsh and
=(...) because I can't be bothered to do the whole mktemp + trap dance),
and sic vim on it, giving vim the TTY for its standard input so it
doesn't throw a fit.